requirements.txt
django==4.1.7
Pillow==22.3.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-filter==22.1
Django-Verify-Email==2.0.3